+++ packages/hydrogen/src/client.ts
@@ -3,5 +3,5 @@ export * from './hooks';
 export * from './foundation/useServerState';
 export * from './foundation/useShop';
 export * from './foundation/ServerStateProvider';
-export {Helmet} from './foundation/Helmet';
+export {Head} from './foundation/Head';
 export * from './utilities';