+++ packages/hydrogen/src/components/Seo/tests/HomePageSeo.test.tsx
@@ -7,7 +7,7 @@ import {TitleSeo} from '../TitleSeo.client';
 import {DescriptionSeo} from '../DescriptionSeo.client';
 
 jest.mock('../../../client', () => ({
-  Helmet({children}) {
+  Head({children}) {
     return children;
   },
 }));